+++ CHANGELOG.md
@@ -1,10 +1,25 @@
 # Change log
 
 ## 1.4.0
-- Add `OUTPUT`, `RESULT`, `PASS_COUNT` and `FAIL_COUNT` constants.
 - Add printing error codes for invalid commands.
 - Add `Test` command.
 - Add operators for strings.
+- Add printing color. Allows for foreground and background. Allows for common word colors. Ex. `Print "text" in red.` or `Print "text" in red with white.`
+  - Red
+  - Yellow
+  - Green
+  - Blue
+  - Magenta
+  - Cyan
+  - White
+  - Black
+- Changed `Print` and `Show` to only allow for string or variable types.
+- Add `Input` command. Accepts input from the user.
+- Add `Countdown` command. Waits, but shows the time remaining in real time.
+- Add `Exit` command. Quits the program.
+- Add `and` and `or` operators.
+- Add `OUTPUT`, `RESULT`, `PASS_COUNT`, `FAIL_COUNT`, `true`, `false` constants.
+- Fix various bugs.
 
 ## 1.3.1
 - Update installer to ignore extra dll files.